+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The CHANGELOG is available on [Github](https://github.com/luc-tielen/souffle-has
 ### Added
 
 - WebAssembly support
+- The `string` data type to the language
 - Wildcards are now supported in rule bodies
 - Assignments are now supported in rule bodies
 - Support for multiple occurences of the same variable in a single clause of